revamp of hadling of parameters
This implementation simplifies how the user interacts with setting the
parameters. The user, in the vast majority of cases, will only need to
interact with `Study.params` and the StudyParam class wil handle the heavy
lifting. Whereas, previously the user would have to set the values in 3 dicts
`madx_params`, `sixtrack_params` and `oneturn_sixtrack_params` eventhough alot
of the values should be identical accross these 3 dicts. I also moved alot of
the default parameter values in machineparams.py.
These default parameters are now simply stored in dictionnaries which
makes it very easy for the user to add his own default configurations.
StudyParam also has a more elegant way of doing pre-calculations, using a
calculation queue. I added a unit test for StudyParam which shows off
most of it's functionnality.