rework of jobs
the job executables have been been refactored into classes, with the common methods in a base class in lib/sixtrack_job.py. I changed the sixtrack_input field from list (which was needed when there used to be multiple fort mother files) to string so that it is consistent with the madx_input field, which simplifies a few things in study.py and in the jobs. I made some slight changes to the SixDB to allow for opening and closing a database whithout having to recreate a new instance, which is useful for the jobs.